The Pedixplorer package for pedigree data

The Pedixplorer package for pedigree data an updated package of the kinship2 package. The kinship2 package was originally written by Terry Therneau and Jason Sinnwell. The Pedixplorer package is a fork of the kinship2 package with additional functionality and bug fixes.

Details

The package download, NEWS, and README are available on CRAN: Kinship2 for the previous version of the package.

Functions

Below are listed some of the most widely used functions available in arsenal:

Pedigree(): Contstructor of the Pedigree class, given identifiers, sex, affection status(es), and special relationships

kinship(): Calculates the kinship matrix, the probability having an allele sampled from two individuals be the same via IBD.

plot() : Method to transform a Pedigree object into a graphical plot. Allows extra information to be included in the id under the plot symbol. This method use the plot_fromdf() function to transform the Pedigree object into a data frame of graphical elements, the same is done for the legend with the ped_to_legdf() function. When done, the data frames are plotted with the plot_fromdf() function.

shrink(): Shrink a Pedigree to a specific bit size, removing non-informative members first.

bit_size(): Approximate the output from SAS's PROC FREQ procedure when using the /list option of the TABLE statement.

Data

sampleped(): Pedigree example data sets with two pedigrees

minnbreast(): Larger cohort of pedigrees from MN breast cancer study

See also

Author

Maintainer: Louis Le Nézet louislenezet@gmail.com (ORCID) [contributor]

Authors:

Jason Sinnwell sinnwell.jason@mayo.edu

Terry Therneau

Other contributors:

Daniel Schaid [contributor]

Elizabeth Atkinson [contributor]

Examples

library(Pedixplorer)