Circular element

Create a list of x and y coordinates for a circle with a given number of slices.

circfun(nslice, n = 50)

Arguments

nslice

Number of slices in the circle

n

Total number of points in the circle

Value

A list of x and y coordinates per slice.

Examples

circfun(1)
#> [[1]]
#> [[1]]$x
#>  [1]  0.000000e+00  3.061617e-17 -6.393858e-02 -1.268273e-01 -1.876335e-01
#>  [6] -2.453588e-01 -2.990553e-01 -3.478413e-01 -3.909157e-01 -4.275714e-01
#> [11] -4.572063e-01 -4.793339e-01 -4.935909e-01 -4.997431e-01 -4.976896e-01
#> [16] -4.874640e-01 -4.692342e-01 -4.432997e-01 -4.100861e-01 -3.701390e-01
#> [21] -3.241142e-01 -2.727675e-01 -2.169419e-01 -1.575541e-01 -9.557931e-02
#> [26] -3.203511e-02  3.203511e-02  9.557931e-02  1.575541e-01  2.169419e-01
#> [31]  2.727675e-01  3.241142e-01  3.701390e-01  4.100861e-01  4.432997e-01
#> [36]  4.692342e-01  4.874640e-01  4.976896e-01  4.997431e-01  4.935909e-01
#> [41]  4.793339e-01  4.572063e-01  4.275714e-01  3.909157e-01  3.478413e-01
#> [46]  2.990553e-01  2.453588e-01  1.876335e-01  1.268273e-01  6.393858e-02
#> [51]  1.530808e-16
#> 
#> [[1]]$y
#>  [1] 0.500000000 0.000000000 0.004104993 0.016352568 0.036541621 0.064340648
#>  [7] 0.099293189 0.140825325 0.188255099 0.240803716 0.297608328 0.357736207
#> [13] 0.420200052 0.483974211 0.548011513 0.611260467 0.672682527 0.731269145
#> [19] 0.786058330 0.836150445 0.880722979 0.919044052 0.950484434 0.974527874
#> [25] 0.990779578 0.998972696 0.998972696 0.990779578 0.974527874 0.950484434
#> [31] 0.919044052 0.880722979 0.836150445 0.786058330 0.731269145 0.672682527
#> [37] 0.611260467 0.548011513 0.483974211 0.420200052 0.357736207 0.297608328
#> [43] 0.240803716 0.188255099 0.140825325 0.099293189 0.064340648 0.036541621
#> [49] 0.016352568 0.004104993 0.000000000
#> 
#> 
circfun(1, 10)
#> [[1]]
#> [[1]]$x
#>  [1]  0.000000e+00  3.061617e-17 -3.213938e-01 -4.924039e-01 -4.330127e-01
#>  [6] -1.710101e-01  1.710101e-01  4.330127e-01  4.924039e-01  3.213938e-01
#> [11]  1.530808e-16
#> 
#> [[1]]$y
#>  [1] 0.5000000 0.0000000 0.1169778 0.4131759 0.7500000 0.9698463 0.9698463
#>  [8] 0.7500000 0.4131759 0.1169778 0.0000000
#> 
#> 
circfun(4, 50)
#> [[1]]
#> [[1]]$x
#>  [1]  0.000000e+00  3.061617e-17 -6.526310e-02 -1.294095e-01 -1.913417e-01
#>  [6] -2.500000e-01 -3.043807e-01 -3.535534e-01 -3.966767e-01 -4.330127e-01
#> [11] -4.619398e-01 -4.829629e-01 -4.957224e-01 -5.000000e-01
#> 
#> [[1]]$y
#>  [1] 0.500000000 0.000000000 0.004277569 0.017037087 0.038060234 0.066987298
#>  [7] 0.103323330 0.146446609 0.195619285 0.250000000 0.308658284 0.370590477
#> [13] 0.434736904 0.500000000
#> 
#> 
#> [[2]]
#> [[2]]$x
#>  [1]  0.000000e+00 -5.000000e-01 -4.957224e-01 -4.829629e-01 -4.619398e-01
#>  [6] -4.330127e-01 -3.966767e-01 -3.535534e-01 -3.043807e-01 -2.500000e-01
#> [11] -1.913417e-01 -1.294095e-01 -6.526310e-02 -9.184851e-17
#> 
#> [[2]]$y
#>  [1] 0.5000000 0.5000000 0.5652631 0.6294095 0.6913417 0.7500000 0.8043807
#>  [8] 0.8535534 0.8966767 0.9330127 0.9619398 0.9829629 0.9957224 1.0000000
#> 
#> 
#> [[3]]
#> [[3]]$x
#>  [1]  0.000000e+00 -9.184851e-17  6.526310e-02  1.294095e-01  1.913417e-01
#>  [6]  2.500000e-01  3.043807e-01  3.535534e-01  3.966767e-01  4.330127e-01
#> [11]  4.619398e-01  4.829629e-01  4.957224e-01  5.000000e-01
#> 
#> [[3]]$y
#>  [1] 0.5000000 1.0000000 0.9957224 0.9829629 0.9619398 0.9330127 0.8966767
#>  [8] 0.8535534 0.8043807 0.7500000 0.6913417 0.6294095 0.5652631 0.5000000
#> 
#> 
#> [[4]]
#> [[4]]$x
#>  [1] 0.000000e+00 5.000000e-01 4.957224e-01 4.829629e-01 4.619398e-01
#>  [6] 4.330127e-01 3.966767e-01 3.535534e-01 3.043807e-01 2.500000e-01
#> [11] 1.913417e-01 1.294095e-01 6.526310e-02 1.530808e-16
#> 
#> [[4]]$y
#>  [1] 0.500000000 0.500000000 0.434736904 0.370590477 0.308658284 0.250000000
#>  [7] 0.195619285 0.146446609 0.103323330 0.066987298 0.038060234 0.017037087
#> [13] 0.004277569 0.000000000
#> 
#>