Descendants of individuals

Find all the descendants of a particular list of individuals given a Pedigree object.

Usage

# S4 method for class 'character_OR_integer,character_OR_integer'
descendants(idlist, obj, dadid, momid)

# S4 method for class 'character_OR_integer,Pedigree'
descendants(idlist, obj)

# S4 method for class 'character_OR_integer,Ped'
descendants(idlist, obj)

Arguments

idlist

List of individuals identifiers to be considered

obj

A Ped or Pedigree object or a vector of the individuals identifiers.

dadid

A vector containing for each subject, the identifiers of the biologicals fathers.

momid

A vector containing for each subject, the identifiers of the biologicals mothers.

Value

Vector of all descendants of the individuals in idlist. The list is not ordered.

Examples

data("sampleped")
pedi <- Pedigree(sampleped)
#>    id dadid momid famid 
#> "101"    NA    NA   "1" 
#>    id dadid momid famid 
#> "102"    NA    NA   "1" 
#>    id dadid momid famid 
#> "103" "135" "136"   "1" 
#>    id dadid momid famid 
#> "104"    NA    NA   "1" 
#>    id dadid momid famid 
#> "105"    NA    NA   "1" 
#>    id dadid momid famid 
#> "106"    NA    NA   "1" 
#>    id dadid momid famid 
#> "107"    NA    NA   "1" 
#>    id dadid momid famid 
#> "108"    NA    NA   "1" 
#>    id dadid momid famid 
#> "109" "101" "102"   "1" 
#>    id dadid momid famid 
#> "110" "103" "104"   "1" 
#>    id dadid momid famid 
#> "111" "103" "104"   "1" 
#>    id dadid momid famid 
#> "112" "103" "104"   "1" 
#>    id dadid momid famid 
#> "113"    NA    NA   "1" 
#>    id dadid momid famid 
#> "114" "103" "104"   "1" 
#>    id dadid momid famid 
#> "115" "105" "106"   "1" 
#>    id dadid momid famid 
#> "116" "105" "106"   "1" 
#>    id dadid momid famid 
#> "117"    NA    NA   "1" 
#>    id dadid momid famid 
#> "118" "105" "106"   "1" 
#>    id dadid momid famid 
#> "119" "105" "106"   "1" 
#>    id dadid momid famid 
#> "120" "107" "108"   "1" 
#>    id dadid momid famid 
#> "121" "110" "109"   "1" 
#>    id dadid momid famid 
#> "122" "110" "109"   "1" 
#>    id dadid momid famid 
#> "123" "110" "109"   "1" 
#>    id dadid momid famid 
#> "124" "110" "109"   "1" 
#>    id dadid momid famid 
#> "125" "112" "118"   "1" 
#>    id dadid momid famid 
#> "126" "112" "118"   "1" 
#>    id dadid momid famid 
#> "127" "114" "115"   "1" 
#>    id dadid momid famid 
#> "128" "114" "115"   "1" 
#>    id dadid momid famid 
#> "129" "117" "116"   "1" 
#>    id dadid momid famid 
#> "130" "119" "120"   "1" 
#>    id dadid momid famid 
#> "131" "119" "120"   "1" 
#>    id dadid momid famid 
#> "132" "119" "120"   "1" 
#>    id dadid momid famid 
#> "133" "119" "120"   "1" 
#>    id dadid momid famid 
#> "134" "119" "120"   "1" 
#>    id dadid momid famid 
#> "135"    NA    NA   "1" 
#>    id dadid momid famid 
#> "136"    NA    NA   "1" 
#>    id dadid momid famid 
#> "137"    NA    NA   "1" 
#>    id dadid momid famid 
#> "138" "135" "136"   "1" 
#>    id dadid momid famid 
#> "139" "137" "138"   "1" 
#>    id dadid momid famid 
#> "140" "137" "138"   "1" 
#>    id dadid momid famid 
#> "141" "137" "138"   "1" 
#>    id dadid momid famid 
#> "201"    NA    NA   "2" 
#>    id dadid momid famid 
#> "202"    NA    NA   "2" 
#>    id dadid momid famid 
#> "203"    NA    NA   "2" 
#>    id dadid momid famid 
#> "204" "201" "202"   "2" 
#>    id dadid momid famid 
#> "205" "201" "202"   "2" 
#>    id dadid momid famid 
#> "206" "201" "202"   "2" 
#>    id dadid momid famid 
#> "207" "201" "202"   "2" 
#>    id dadid momid famid 
#> "208" "201" "202"   "2" 
#>    id dadid momid famid 
#> "209"    NA    NA   "2" 
#>    id dadid momid famid 
#> "210" "203" "204"   "2" 
#>    id dadid momid famid 
#> "211" "203" "204"   "2" 
#>    id dadid momid famid 
#> "212" "209" "208"   "2" 
#>    id dadid momid famid 
#> "213" "209" "208"   "2" 
#>    id dadid momid famid 
#> "214" "209" "208"   "2" 
descendants(c("1_101", "2_208"), pedi)
#> [1] "1_109" "2_212" "2_213" "2_214" "1_121" "1_122" "1_123" "1_124"