Get twin relationships

Get twin relationships

Usage

get_twin_rel(obj)

Arguments

obj

A Pedigree object

Value

A list containing components

twinset the set of twins

twinrel the twins relationships

twinord the order of the twins

Details

This routine function determine the twin relationships in a Pedigree. It determine the order of the twins in the Pedigree. It is used by auto_hint().

See also