Shiny module to export plot

This module allow to export multiple type of plot from a reactive object. The file type currently supported are png, pdf and html. The function is composed of two parts: the UI and the server. The UI is called with the function plot_download_ui() and the server with the function plot_download_server().

Usage

plot_download_ui(id)

plot_download_server(
  id,
  my_plot,
  filename = "saveplot",
  label = "Download",
  width = 500,
  height = 500,
  ext = "png"
)

plot_download_demo()

Arguments

id

A string.

my_plot

Reactive object containing the plot.

filename

A string to name the file.

label

A string to name the download button.

width

A numeric to set the width of the plot.

height

A numeric to set the height of the plot.

ext

A string to set the extension of the file.

Value

A shiny module to export a plot.

Examples

if (interactive()) {
    plot_download_demo()
}