Create a plot from a data.frame

This function is used to create a plot from a data.frame.

If ggplot_gen = TRUE, the plot will be generated with ggplot2 and will be returned invisibly.

Usage

plot_fromdf(
  df,
  usr = NULL,
  title = NULL,
  ggplot_gen = FALSE,
  boxw = 1,
  boxh = 1,
  add_to_existing = FALSE,
  title_cex = 2
)

Arguments

df

A data.frame with the following columns:

type: The type of element to plot. Can be text, segments, arc or other polygons. For polygons, the name of the polygon must be in the form poly_*_* where poly is one of the type given by polygons(), the first * is the number of slice in the polygon and the second * is the position of the division of the polygon.

x0: The x coordinate of the center of the element.

y0: The y coordinate of the center of the element.

x1: The x coordinate of the end of the element. Only used for segments and arc.

y1: The y coordinate of the end of the element. Only used for segments and arc.

fill: The fill color of the element.

border: The border color of the element.

density: The density of the element.

angle: The angle of the element.

label: The label of the element. Only used for text.

cex: The size of the element.

adjx: The x adjustment of the element. Only used for text.

adjy: The y adjustment of the element. Only used for text.

usr

The user coordinates of the plot.

title

The title of the plot.

ggplot_gen

If TRUE add the segments to the ggplot object

boxw

Width of the polygons elements

boxh

Height of the polygons elements

add_to_existing

If TRUE, the plot will be added to the current plot.

title_cex

The size of the title.

Value

an invisible ggplot object and a plot on the current plotting device

Examples

data(sampleped)
ped1 <- Pedigree(sampleped[sampleped$famid == 1,])
#>    id dadid momid famid 
#> "101"    NA    NA   "1" 
#>    id dadid momid famid 
#> "102"    NA    NA   "1" 
#>    id dadid momid famid 
#> "103" "135" "136"   "1" 
#>    id dadid momid famid 
#> "104"    NA    NA   "1" 
#>    id dadid momid famid 
#> "105"    NA    NA   "1" 
#>    id dadid momid famid 
#> "106"    NA    NA   "1" 
#>    id dadid momid famid 
#> "107"    NA    NA   "1" 
#>    id dadid momid famid 
#> "108"    NA    NA   "1" 
#>    id dadid momid famid 
#> "109" "101" "102"   "1" 
#>    id dadid momid famid 
#> "110" "103" "104"   "1" 
#>    id dadid momid famid 
#> "111" "103" "104"   "1" 
#>    id dadid momid famid 
#> "112" "103" "104"   "1" 
#>    id dadid momid famid 
#> "113"    NA    NA   "1" 
#>    id dadid momid famid 
#> "114" "103" "104"   "1" 
#>    id dadid momid famid 
#> "115" "105" "106"   "1" 
#>    id dadid momid famid 
#> "116" "105" "106"   "1" 
#>    id dadid momid famid 
#> "117"    NA    NA   "1" 
#>    id dadid momid famid 
#> "118" "105" "106"   "1" 
#>    id dadid momid famid 
#> "119" "105" "106"   "1" 
#>    id dadid momid famid 
#> "120" "107" "108"   "1" 
#>    id dadid momid famid 
#> "121" "110" "109"   "1" 
#>    id dadid momid famid 
#> "122" "110" "109"   "1" 
#>    id dadid momid famid 
#> "123" "110" "109"   "1" 
#>    id dadid momid famid 
#> "124" "110" "109"   "1" 
#>    id dadid momid famid 
#> "125" "112" "118"   "1" 
#>    id dadid momid famid 
#> "126" "112" "118"   "1" 
#>    id dadid momid famid 
#> "127" "114" "115"   "1" 
#>    id dadid momid famid 
#> "128" "114" "115"   "1" 
#>    id dadid momid famid 
#> "129" "117" "116"   "1" 
#>    id dadid momid famid 
#> "130" "119" "120"   "1" 
#>    id dadid momid famid 
#> "131" "119" "120"   "1" 
#>    id dadid momid famid 
#> "132" "119" "120"   "1" 
#>    id dadid momid famid 
#> "133" "119" "120"   "1" 
#>    id dadid momid famid 
#> "134" "119" "120"   "1" 
#>    id dadid momid famid 
#> "135"    NA    NA   "1" 
#>    id dadid momid famid 
#> "136"    NA    NA   "1" 
#>    id dadid momid famid 
#> "137"    NA    NA   "1" 
#>    id dadid momid famid 
#> "138" "135" "136"   "1" 
#>    id dadid momid famid 
#> "139" "137" "138"   "1" 
#>    id dadid momid famid 
#> "140" "137" "138"   "1" 
#>    id dadid momid famid 
#> "141" "137" "138"   "1" 
lst <- ped_to_plotdf(ped1)
#> Individuals:  1_113 won't be plotted
if (interactive()) {
    plot_fromdf(lst$df, lst$par_usr$usr,
        boxw = lst$par_usr$boxw, boxh = lst$par_usr$boxh
    )
}