Rescale continuous vector to have specified minimum and maximum

Rescale continuous vector to have specified minimum and maximum

Usage

rescale(x, to = c(0, 1), from = range(x, na.rm = TRUE, finite = TRUE))

Arguments

x

continuous vector of values to manipulate.

to

output range (numeric vector of length two)

from

input range (vector of length two). If not given, is calculated from the range of x

Value

A numeric vector of the same length as x with values rescaled to the range specified by to.

Details

Objects of class <AsIs> are returned unaltered.

Examples

rescale(1:100)
#>   [1] 0.00000000 0.01010101 0.02020202 0.03030303 0.04040404 0.05050505
#>   [7] 0.06060606 0.07070707 0.08080808 0.09090909 0.10101010 0.11111111
#>  [13] 0.12121212 0.13131313 0.14141414 0.15151515 0.16161616 0.17171717
#>  [19] 0.18181818 0.19191919 0.20202020 0.21212121 0.22222222 0.23232323
#>  [25] 0.24242424 0.25252525 0.26262626 0.27272727 0.28282828 0.29292929
#>  [31] 0.30303030 0.31313131 0.32323232 0.33333333 0.34343434 0.35353535
#>  [37] 0.36363636 0.37373737 0.38383838 0.39393939 0.40404040 0.41414141
#>  [43] 0.42424242 0.43434343 0.44444444 0.45454545 0.46464646 0.47474747
#>  [49] 0.48484848 0.49494949 0.50505051 0.51515152 0.52525253 0.53535354
#>  [55] 0.54545455 0.55555556 0.56565657 0.57575758 0.58585859 0.59595960
#>  [61] 0.60606061 0.61616162 0.62626263 0.63636364 0.64646465 0.65656566
#>  [67] 0.66666667 0.67676768 0.68686869 0.69696970 0.70707071 0.71717172
#>  [73] 0.72727273 0.73737374 0.74747475 0.75757576 0.76767677 0.77777778
#>  [79] 0.78787879 0.79797980 0.80808081 0.81818182 0.82828283 0.83838384
#>  [85] 0.84848485 0.85858586 0.86868687 0.87878788 0.88888889 0.89898990
#>  [91] 0.90909091 0.91919192 0.92929293 0.93939394 0.94949495 0.95959596
#>  [97] 0.96969697 0.97979798 0.98989899 1.00000000
rescale(runif(50))
#>  [1] 0.37256665 0.98419514 0.26958502 0.67447864 0.73382229 0.17168265
#>  [7] 0.98939778 0.74028613 0.02106965 0.52005368 0.69265871 0.68508392
#> [13] 0.00000000 0.20253867 0.28098545 0.63079380 0.46670415 0.41787229
#> [19] 0.70371672 0.95608491 0.15540515 0.19351020 0.67633640 0.48736223
#> [25] 0.63622771 0.65561838 0.06753001 0.76538158 0.76962829 1.00000000
#> [31] 0.97895593 0.37306843 0.44811278 0.29600496 0.14950314 0.52147223
#> [37] 0.48193369 0.77966894 0.18025145 0.71097422 0.03542097 0.33661546
#> [43] 0.82749766 0.25283218 0.56156826 0.31734703 0.58889325 0.16705650
#> [49] 0.95523796 0.53283969
rescale(1)
#> [1] NaN