Set plotting area

Set plotting area

set_plot_area(cex, id, maxlev, xrange, symbolsize, precision = 3, ...)

Arguments

cex

Character expansion of the text

id

A character vector with the identifiers of each individuals

maxlev

Maximum level

xrange

Range of x values

symbolsize

Size of the symbols

precision

The number of significant digits to round the solution to.

...

Other arguments passed to par()

Value

List of user coordinates, old par, box width, box height, label height and leg height