Gender variable to ordered factor

Gender variable to ordered factor

Usage

sex_to_factor(sex)

Arguments

sex

A character, factor or numeric vector corresponding to the gender of the individuals. This will be transformed to an ordered factor with the following levels: male < female < unknown < terminated The following values are recognized:

character() or factor() : "f", "m", "woman", "man", "male", "female", "unknown", "terminated"

numeric() : 1 = "male", 2 = "female", 3 = "unknown", 4 = "terminated"

Value

an ordered factor vector containing the transformed variable "male" < "female" < "unknown" < "terminated"

Examples

sex_to_factor(c(1, 2, 3, 4, "f", "m", "man", "female"))
#> [1] male       female     unknown    terminated female     male       male      
#> [8] female    
#> Levels: male < female < unknown < terminated