Circular element

Create a list of x and y coordinates for a circle with a given number of slices.

Usage

circfun(nslice, n = 50, start = 0)

Arguments

nslice

Number of slices in the circle

n

Total number of points in the circle

Value

A list of x and y coordinates per slice.

Examples

Pedixplorer:::circfun(1)
#> [[1]]
#> [[1]]$x
#>  [1]  0.50000000  0.49589501  0.48364743  0.46345838  0.43565935  0.40070681
#>  [7]  0.35917468  0.31174490  0.25919628  0.20239167  0.14226379  0.07979995
#> [13]  0.01602579 -0.04801151 -0.11126047 -0.17268253 -0.23126915 -0.28605833
#> [19] -0.33615045 -0.38072298 -0.41904405 -0.45048443 -0.47452787 -0.49077958
#> [25] -0.49897270 -0.49897270 -0.49077958 -0.47452787 -0.45048443 -0.41904405
#> [31] -0.38072298 -0.33615045 -0.28605833 -0.23126915 -0.17268253 -0.11126047
#> [37] -0.04801151  0.01602579  0.07979995  0.14226379  0.20239167  0.25919628
#> [43]  0.31174490  0.35917468  0.40070681  0.43565935  0.46345838  0.48364743
#> [49]  0.49589501  0.50000000
#> 
#> [[1]]$y
#>  [1]  0.000000e+00  6.393858e-02  1.268273e-01  1.876335e-01  2.453588e-01
#>  [6]  2.990553e-01  3.478413e-01  3.909157e-01  4.275714e-01  4.572063e-01
#> [11]  4.793339e-01  4.935909e-01  4.997431e-01  4.976896e-01  4.874640e-01
#> [16]  4.692342e-01  4.432997e-01  4.100861e-01  3.701390e-01  3.241142e-01
#> [21]  2.727675e-01  2.169419e-01  1.575541e-01  9.557931e-02  3.203511e-02
#> [26] -3.203511e-02 -9.557931e-02 -1.575541e-01 -2.169419e-01 -2.727675e-01
#> [31] -3.241142e-01 -3.701390e-01 -4.100861e-01 -4.432997e-01 -4.692342e-01
#> [36] -4.874640e-01 -4.976896e-01 -4.997431e-01 -4.935909e-01 -4.793339e-01
#> [41] -4.572063e-01 -4.275714e-01 -3.909157e-01 -3.478413e-01 -2.990553e-01
#> [46] -2.453588e-01 -1.876335e-01 -1.268273e-01 -6.393858e-02 -1.224647e-16
#> 
#> 
Pedixplorer:::circfun(1, 10)
#> [[1]]
#> [[1]]$x
#>  [1]  0.50000000  0.38302222  0.08682409 -0.25000000 -0.46984631 -0.46984631
#>  [7] -0.25000000  0.08682409  0.38302222  0.50000000
#> 
#> [[1]]$y
#>  [1]  0.000000e+00  3.213938e-01  4.924039e-01  4.330127e-01  1.710101e-01
#>  [6] -1.710101e-01 -4.330127e-01 -4.924039e-01 -3.213938e-01 -1.224647e-16
#> 
#> 
Pedixplorer:::circfun(4, 50)
#> [[1]]
#> [[1]]$x
#>  [1] 0.000000e+00 5.000000e-01 4.957224e-01 4.829629e-01 4.619398e-01
#>  [6] 4.330127e-01 3.966767e-01 3.535534e-01 3.043807e-01 2.500000e-01
#> [11] 1.913417e-01 1.294095e-01 6.526310e-02 3.061617e-17
#> 
#> [[1]]$y
#>  [1] 0.0000000 0.0000000 0.0652631 0.1294095 0.1913417 0.2500000 0.3043807
#>  [8] 0.3535534 0.3966767 0.4330127 0.4619398 0.4829629 0.4957224 0.5000000
#> 
#> 
#> [[2]]
#> [[2]]$x
#>  [1]  0.000000e+00  3.061617e-17 -6.526310e-02 -1.294095e-01 -1.913417e-01
#>  [6] -2.500000e-01 -3.043807e-01 -3.535534e-01 -3.966767e-01 -4.330127e-01
#> [11] -4.619398e-01 -4.829629e-01 -4.957224e-01 -5.000000e-01
#> 
#> [[2]]$y
#>  [1] 0.000000e+00 5.000000e-01 4.957224e-01 4.829629e-01 4.619398e-01
#>  [6] 4.330127e-01 3.966767e-01 3.535534e-01 3.043807e-01 2.500000e-01
#> [11] 1.913417e-01 1.294095e-01 6.526310e-02 6.123234e-17
#> 
#> 
#> [[3]]
#> [[3]]$x
#>  [1]  0.000000e+00 -5.000000e-01 -4.957224e-01 -4.829629e-01 -4.619398e-01
#>  [6] -4.330127e-01 -3.966767e-01 -3.535534e-01 -3.043807e-01 -2.500000e-01
#> [11] -1.913417e-01 -1.294095e-01 -6.526310e-02 -9.184851e-17
#> 
#> [[3]]$y
#>  [1]  0.000000e+00  6.123234e-17 -6.526310e-02 -1.294095e-01 -1.913417e-01
#>  [6] -2.500000e-01 -3.043807e-01 -3.535534e-01 -3.966767e-01 -4.330127e-01
#> [11] -4.619398e-01 -4.829629e-01 -4.957224e-01 -5.000000e-01
#> 
#> 
#> [[4]]
#> [[4]]$x
#>  [1]  0.000000e+00 -9.184851e-17  6.526310e-02  1.294095e-01  1.913417e-01
#>  [6]  2.500000e-01  3.043807e-01  3.535534e-01  3.966767e-01  4.330127e-01
#> [11]  4.619398e-01  4.829629e-01  4.957224e-01  5.000000e-01
#> 
#> [[4]]$y
#>  [1]  0.000000e+00 -5.000000e-01 -4.957224e-01 -4.829629e-01 -4.619398e-01
#>  [6] -4.330127e-01 -3.966767e-01 -3.535534e-01 -3.043807e-01 -2.500000e-01
#> [11] -1.913417e-01 -1.294095e-01 -6.526310e-02 -1.224647e-16
#> 
#>