Kinship matrix

Compute the kinship matrix for a set of related autosomal subjects. The function is generic, and can accept a Pedigree, a Ped or a vector as the first argument.

Usage

# S4 method for class 'Ped'
kinship(obj, chrtype = "autosome")

# S4 method for class 'character'
kinship(obj, dadid, momid, sex, chrtype = "autosome")

# S4 method for class 'Pedigree'
kinship(obj, chrtype = "autosome")

Arguments

obj

A Pedigree or Ped object or a vector of subject identifiers.

chrtype

chromosome type. The currently supported types are 'autosome' and 'X' or 'x'.

dadid

A vector containing for each subject, the identifiers of the biologicals fathers.

momid

A vector containing for each subject, the identifiers of the biologicals mothers.

sex

A character, factor or numeric vector corresponding to the gender of the individuals. This will be transformed to an ordered factor with the following levels: male < female < unknown < terminated The following values are recognized:

character() or factor() : "f", "m", "woman", "man", "male", "female", "unknown", "terminated"

numeric() : 1 = "male", 2 = "female", 3 = "unknown", 4 = "terminated"

Value

When obj is a vector

A matrix of kinship coefficients.

When obj is a Pedigree

A matrix of kinship coefficients ordered by families present in the Pedigree object.

Details

The function will usually be called with a Pedigree. The call with a Ped or a vector is provided for backwards compatibility with an earlier release of the library that was less capable. Note that when using with a Ped or a vector, any information on twins is not available to the function.

When called with a Pedigree, the routine will create a block-diagonal-symmetric sparse matrix object of class dsCMatrix. Since the [i, j] value of the result is 0 for any two unrelated individuals i and j and a Matrix utilizes sparse representation, the resulting object is often orders of magnitude smaller than an ordinary matrix.

Two genes G1 and G2 are identical by descent (IBD) if they are both physical copies of the same ancestral gene; two genes are identical by state if they represent the same allele. So the brown eye gene that I inherited from my mother is IBD with hers; the same gene in an unrelated individual is not.

The kinship coefficient between two subjects is the probability that a randomly selected allele from a locus will be IBD between them. It is obviously 0 between unrelated individuals. For an autosomal site and no inbreeding it will be 0.5 for an individual with themselves, .25 between mother and child, .125 between an uncle and neice, etc.

The computation is based on a recursive algorithm described in Lange, which assumes that the founder alleles are all independent.

References

K Lange, Mathematical and Statistical Methods for Genetic Analysis, Springer-Verlag, New York, 1997.

See also

Examples

kinship(c("A", "B", "C", "D", "E"), c("C", "D", "0", "0", "0"),
    c("E", "E", "0", "0", "0"), sex = c(1, 2, 1, 2, 1))
#> 5 x 5 sparse Matrix of class "dsCMatrix"
#>       A     B    C    D    E
#> A 0.500 0.125 0.25 .    0.25
#> B 0.125 0.500 .    0.25 0.25
#> C 0.250 .     0.50 .    .   
#> D .     0.250 .    0.50 .   
#> E 0.250 0.250 .    .    0.50
kinship(c("A", "B", "C", "D", "E"), c("C", "D", "0", "0", "0"),
    c("E", "E", "0", "0", "0"), sex = c(1, 2, 1, 2, 1),
    chrtype = "x"
)
#> 5 x 5 sparse Matrix of class "dsCMatrix"
#>     A    B C    D   E
#> A 1.0 0.50 . .    1.0
#> B 0.5 0.50 . 0.25 0.5
#> C .   .    1 .    .  
#> D .   0.25 . 0.50 .  
#> E 1.0 0.50 . .    1.0

data(sampleped)
ped <- Pedigree(sampleped)
kinship(ped)
#> 55 x 55 sparse Matrix of class "dsCMatrix"
#>   [[ suppressing 55 column names ‘1_101’, ‘1_102’, ‘1_103’ ... ]]
#>                                                                                
#> 1_101 0.500 .     .     .     .     .     .     .     0.25 .      .      .     
#> 1_102 .     0.500 .     .     .     .     .     .     0.25 .      .      .     
#> 1_103 .     .     0.500 .     .     .     .     .     .    0.2500 0.2500 0.2500
#> 1_104 .     .     .     0.500 .     .     .     .     .    0.2500 0.2500 0.2500
#> 1_105 .     .     .     .     0.500 .     .     .     .    .      .      .     
#> 1_106 .     .     .     .     .     0.500 .     .     .    .      .      .     
#> 1_107 .     .     .     .     .     .     0.500 .     .    .      .      .     
#> 1_108 .     .     .     .     .     .     .     0.500 .    .      .      .     
#> 1_109 0.250 0.250 .     .     .     .     .     .     0.50 .      .      .     
#> 1_110 .     .     0.250 0.250 .     .     .     .     .    0.5000 0.2500 0.2500
#> 1_111 .     .     0.250 0.250 .     .     .     .     .    0.2500 0.5000 0.2500
#> 1_112 .     .     0.250 0.250 .     .     .     .     .    0.2500 0.2500 0.5000
#> 1_113 .     .     .     .     .     .     .     .     .    .      .      .     
#> 1_114 .     .     0.250 0.250 .     .     .     .     .    0.2500 0.2500 0.2500
#> 1_115 .     .     .     .     0.250 0.250 .     .     .    .      .      .     
#> 1_116 .     .     .     .     0.250 0.250 .     .     .    .      .      .     
#> 1_117 .     .     .     .     .     .     .     .     .    .      .      .     
#> 1_118 .     .     .     .     0.250 0.250 .     .     .    .      .      .     
#> 1_119 .     .     .     .     0.250 0.250 .     .     .    .      .      .     
#> 1_120 .     .     .     .     .     .     0.250 0.250 .    .      .      .     
#> 1_121 0.125 0.125 0.125 0.125 .     .     .     .     0.25 0.2500 0.1250 0.1250
#> 1_122 0.125 0.125 0.125 0.125 .     .     .     .     0.25 0.2500 0.1250 0.1250
#> 1_123 0.125 0.125 0.125 0.125 .     .     .     .     0.25 0.2500 0.1250 0.1250
#> 1_124 0.125 0.125 0.125 0.125 .     .     .     .     0.25 0.2500 0.1250 0.1250
#> 1_125 .     .     0.125 0.125 0.125 0.125 .     .     .    0.1250 0.1250 0.2500
#> 1_126 .     .     0.125 0.125 0.125 0.125 .     .     .    0.1250 0.1250 0.2500
#> 1_127 .     .     0.125 0.125 0.125 0.125 .     .     .    0.1250 0.1250 0.1250
#> 1_128 .     .     0.125 0.125 0.125 0.125 .     .     .    0.1250 0.1250 0.1250
#> 1_129 .     .     .     .     0.125 0.125 .     .     .    .      .      .     
#> 1_130 .     .     .     .     0.125 0.125 0.125 0.125 .    .      .      .     
#> 1_131 .     .     .     .     0.125 0.125 0.125 0.125 .    .      .      .     
#> 1_132 .     .     .     .     0.125 0.125 0.125 0.125 .    .      .      .     
#> 1_133 .     .     .     .     0.125 0.125 0.125 0.125 .    .      .      .     
#> 1_134 .     .     .     .     0.125 0.125 0.125 0.125 .    .      .      .     
#> 1_135 .     .     0.250 .     .     .     .     .     .    0.1250 0.1250 0.1250
#> 1_136 .     .     0.250 .     .     .     .     .     .    0.1250 0.1250 0.1250
#> 1_137 .     .     .     .     .     .     .     .     .    .      .      .     
#> 1_138 .     .     0.250 .     .     .     .     .     .    0.1250 0.1250 0.1250
#> 1_139 .     .     0.125 .     .     .     .     .     .    0.0625 0.0625 0.0625
#> 1_140 .     .     0.125 .     .     .     .     .     .    0.0625 0.0625 0.0625
#> 1_141 .     .     0.125 .     .     .     .     .     .    0.0625 0.0625 0.0625
#> 2_201 .     .     .     .     .     .     .     .     .    .      .      .     
#> 2_202 .     .     .     .     .     .     .     .     .    .      .      .     
#> 2_203 .     .     .     .     .     .     .     .     .    .      .      .     
#> 2_204 .     .     .     .     .     .     .     .     .    .      .      .     
#> 2_205 .     .     .     .     .     .     .     .     .    .      .      .     
#> 2_206 .     .     .     .     .     .     .     .     .    .      .      .     
#> 2_207 .     .     .     .     .     .     .     .     .    .      .      .     
#> 2_208 .     .     .     .     .     .     .     .     .    .      .      .     
#> 2_209 .     .     .     .     .     .     .     .     .    .      .      .     
#> 2_210 .     .     .     .     .     .     .     .     .    .      .      .     
#> 2_211 .     .     .     .     .     .     .     .     .    .      .      .     
#> 2_212 .     .     .     .     .     .     .     .     .    .      .      .     
#> 2_213 .     .     .     .     .     .     .     .     .    .      .      .     
#> 2_214 .     .     .     .     .     .     .     .     .    .      .      .     
#>                                                                           
#> 1_101 .   .      .     .     .    .     .     .    0.12500 0.12500 0.12500
#> 1_102 .   .      .     .     .    .     .     .    0.12500 0.12500 0.12500
#> 1_103 .   0.2500 .     .     .    .     .     .    0.12500 0.12500 0.12500
#> 1_104 .   0.2500 .     .     .    .     .     .    0.12500 0.12500 0.12500
#> 1_105 .   .      0.250 0.250 .    0.250 0.250 .    .       .       .      
#> 1_106 .   .      0.250 0.250 .    0.250 0.250 .    .       .       .      
#> 1_107 .   .      .     .     .    .     .     0.25 .       .       .      
#> 1_108 .   .      .     .     .    .     .     0.25 .       .       .      
#> 1_109 .   .      .     .     .    .     .     .    0.25000 0.25000 0.25000
#> 1_110 .   0.2500 .     .     .    .     .     .    0.25000 0.25000 0.25000
#> 1_111 .   0.2500 .     .     .    .     .     .    0.12500 0.12500 0.12500
#> 1_112 .   0.2500 .     .     .    .     .     .    0.12500 0.12500 0.12500
#> 1_113 0.5 .      .     .     .    .     .     .    .       .       .      
#> 1_114 .   0.5000 .     .     .    .     .     .    0.12500 0.12500 0.12500
#> 1_115 .   .      0.500 0.250 .    0.250 0.250 .    .       .       .      
#> 1_116 .   .      0.250 0.500 .    0.250 0.250 .    .       .       .      
#> 1_117 .   .      .     .     0.50 .     .     .    .       .       .      
#> 1_118 .   .      0.250 0.250 .    0.500 0.250 .    .       .       .      
#> 1_119 .   .      0.250 0.250 .    0.250 0.500 .    .       .       .      
#> 1_120 .   .      .     .     .    .     .     0.50 .       .       .      
#> 1_121 .   0.1250 .     .     .    .     .     .    0.50000 0.25000 0.25000
#> 1_122 .   0.1250 .     .     .    .     .     .    0.25000 0.50000 0.25000
#> 1_123 .   0.1250 .     .     .    .     .     .    0.25000 0.25000 0.50000
#> 1_124 .   0.1250 .     .     .    .     .     .    0.25000 0.25000 0.25000
#> 1_125 .   0.1250 0.125 0.125 .    0.250 0.125 .    0.06250 0.06250 0.06250
#> 1_126 .   0.1250 0.125 0.125 .    0.250 0.125 .    0.06250 0.06250 0.06250
#> 1_127 .   0.2500 0.250 0.125 .    0.125 0.125 .    0.06250 0.06250 0.06250
#> 1_128 .   0.2500 0.250 0.125 .    0.125 0.125 .    0.06250 0.06250 0.06250
#> 1_129 .   .      0.125 0.250 0.25 0.125 0.125 .    .       .       .      
#> 1_130 .   .      0.125 0.125 .    0.125 0.250 0.25 .       .       .      
#> 1_131 .   .      0.125 0.125 .    0.125 0.250 0.25 .       .       .      
#> 1_132 .   .      0.125 0.125 .    0.125 0.250 0.25 .       .       .      
#> 1_133 .   .      0.125 0.125 .    0.125 0.250 0.25 .       .       .      
#> 1_134 .   .      0.125 0.125 .    0.125 0.250 0.25 .       .       .      
#> 1_135 .   0.1250 .     .     .    .     .     .    0.06250 0.06250 0.06250
#> 1_136 .   0.1250 .     .     .    .     .     .    0.06250 0.06250 0.06250
#> 1_137 .   .      .     .     .    .     .     .    .       .       .      
#> 1_138 .   0.1250 .     .     .    .     .     .    0.06250 0.06250 0.06250
#> 1_139 .   0.0625 .     .     .    .     .     .    0.03125 0.03125 0.03125
#> 1_140 .   0.0625 .     .     .    .     .     .    0.03125 0.03125 0.03125
#> 1_141 .   0.0625 .     .     .    .     .     .    0.03125 0.03125 0.03125
#> 2_201 .   .      .     .     .    .     .     .    .       .       .      
#> 2_202 .   .      .     .     .    .     .     .    .       .       .      
#> 2_203 .   .      .     .     .    .     .     .    .       .       .      
#> 2_204 .   .      .     .     .    .     .     .    .       .       .      
#> 2_205 .   .      .     .     .    .     .     .    .       .       .      
#> 2_206 .   .      .     .     .    .     .     .    .       .       .      
#> 2_207 .   .      .     .     .    .     .     .    .       .       .      
#> 2_208 .   .      .     .     .    .     .     .    .       .       .      
#> 2_209 .   .      .     .     .    .     .     .    .       .       .      
#> 2_210 .   .      .     .     .    .     .     .    .       .       .      
#> 2_211 .   .      .     .     .    .     .     .    .       .       .      
#> 2_212 .   .      .     .     .    .     .     .    .       .       .      
#> 2_213 .   .      .     .     .    .     .     .    .       .       .      
#> 2_214 .   .      .     .     .    .     .     .    .       .       .      
#>                                                                          
#> 1_101 0.12500 .       .       .       .       .      .      .      .     
#> 1_102 0.12500 .       .       .       .       .      .      .      .     
#> 1_103 0.12500 0.12500 0.12500 0.12500 0.12500 .      .      .      .     
#> 1_104 0.12500 0.12500 0.12500 0.12500 0.12500 .      .      .      .     
#> 1_105 .       0.12500 0.12500 0.12500 0.12500 0.1250 0.1250 0.1250 0.1250
#> 1_106 .       0.12500 0.12500 0.12500 0.12500 0.1250 0.1250 0.1250 0.1250
#> 1_107 .       .       .       .       .       .      0.1250 0.1250 0.1250
#> 1_108 .       .       .       .       .       .      0.1250 0.1250 0.1250
#> 1_109 0.25000 .       .       .       .       .      .      .      .     
#> 1_110 0.25000 0.12500 0.12500 0.12500 0.12500 .      .      .      .     
#> 1_111 0.12500 0.12500 0.12500 0.12500 0.12500 .      .      .      .     
#> 1_112 0.12500 0.25000 0.25000 0.12500 0.12500 .      .      .      .     
#> 1_113 .       .       .       .       .       .      .      .      .     
#> 1_114 0.12500 0.12500 0.12500 0.25000 0.25000 .      .      .      .     
#> 1_115 .       0.12500 0.12500 0.25000 0.25000 0.1250 0.1250 0.1250 0.1250
#> 1_116 .       0.12500 0.12500 0.12500 0.12500 0.2500 0.1250 0.1250 0.1250
#> 1_117 .       .       .       .       .       0.2500 .      .      .     
#> 1_118 .       0.25000 0.25000 0.12500 0.12500 0.1250 0.1250 0.1250 0.1250
#> 1_119 .       0.12500 0.12500 0.12500 0.12500 0.1250 0.2500 0.2500 0.2500
#> 1_120 .       .       .       .       .       .      0.2500 0.2500 0.2500
#> 1_121 0.25000 0.06250 0.06250 0.06250 0.06250 .      .      .      .     
#> 1_122 0.25000 0.06250 0.06250 0.06250 0.06250 .      .      .      .     
#> 1_123 0.25000 0.06250 0.06250 0.06250 0.06250 .      .      .      .     
#> 1_124 0.50000 0.06250 0.06250 0.06250 0.06250 .      .      .      .     
#> 1_125 0.06250 0.50000 0.25000 0.12500 0.12500 0.0625 0.0625 0.0625 0.0625
#> 1_126 0.06250 0.25000 0.50000 0.12500 0.12500 0.0625 0.0625 0.0625 0.0625
#> 1_127 0.06250 0.12500 0.12500 0.50000 0.25000 0.0625 0.0625 0.0625 0.0625
#> 1_128 0.06250 0.12500 0.12500 0.25000 0.50000 0.0625 0.0625 0.0625 0.0625
#> 1_129 .       0.06250 0.06250 0.06250 0.06250 0.5000 0.0625 0.0625 0.0625
#> 1_130 .       0.06250 0.06250 0.06250 0.06250 0.0625 0.5000 0.2500 0.2500
#> 1_131 .       0.06250 0.06250 0.06250 0.06250 0.0625 0.2500 0.5000 0.2500
#> 1_132 .       0.06250 0.06250 0.06250 0.06250 0.0625 0.2500 0.2500 0.5000
#> 1_133 .       0.06250 0.06250 0.06250 0.06250 0.0625 0.2500 0.2500 0.2500
#> 1_134 .       0.06250 0.06250 0.06250 0.06250 0.0625 0.2500 0.2500 0.2500
#> 1_135 0.06250 0.06250 0.06250 0.06250 0.06250 .      .      .      .     
#> 1_136 0.06250 0.06250 0.06250 0.06250 0.06250 .      .      .      .     
#> 1_137 .       .       .       .       .       .      .      .      .     
#> 1_138 0.06250 0.06250 0.06250 0.06250 0.06250 .      .      .      .     
#> 1_139 0.03125 0.03125 0.03125 0.03125 0.03125 .      .      .      .     
#> 1_140 0.03125 0.03125 0.03125 0.03125 0.03125 .      .      .      .     
#> 1_141 0.03125 0.03125 0.03125 0.03125 0.03125 .      .      .      .     
#> 2_201 .       .       .       .       .       .      .      .      .     
#> 2_202 .       .       .       .       .       .      .      .      .     
#> 2_203 .       .       .       .       .       .      .      .      .     
#> 2_204 .       .       .       .       .       .      .      .      .     
#> 2_205 .       .       .       .       .       .      .      .      .     
#> 2_206 .       .       .       .       .       .      .      .      .     
#> 2_207 .       .       .       .       .       .      .      .      .     
#> 2_208 .       .       .       .       .       .      .      .      .     
#> 2_209 .       .       .       .       .       .      .      .      .     
#> 2_210 .       .       .       .       .       .      .      .      .     
#> 2_211 .       .       .       .       .       .      .      .      .     
#> 2_212 .       .       .       .       .       .      .      .      .     
#> 2_213 .       .       .       .       .       .      .      .      .     
#> 2_214 .       .       .       .       .       .      .      .      .     
#>                                                                            
#> 1_101 .      .      .      .      .    .      .       .       .       .    
#> 1_102 .      .      .      .      .    .      .       .       .       .    
#> 1_103 .      .      0.2500 0.2500 .    0.2500 0.12500 0.12500 0.12500 .    
#> 1_104 .      .      .      .      .    .      .       .       .       .    
#> 1_105 0.1250 0.1250 .      .      .    .      .       .       .       .    
#> 1_106 0.1250 0.1250 .      .      .    .      .       .       .       .    
#> 1_107 0.1250 0.1250 .      .      .    .      .       .       .       .    
#> 1_108 0.1250 0.1250 .      .      .    .      .       .       .       .    
#> 1_109 .      .      .      .      .    .      .       .       .       .    
#> 1_110 .      .      0.1250 0.1250 .    0.1250 0.06250 0.06250 0.06250 .    
#> 1_111 .      .      0.1250 0.1250 .    0.1250 0.06250 0.06250 0.06250 .    
#> 1_112 .      .      0.1250 0.1250 .    0.1250 0.06250 0.06250 0.06250 .    
#> 1_113 .      .      .      .      .    .      .       .       .       .    
#> 1_114 .      .      0.1250 0.1250 .    0.1250 0.06250 0.06250 0.06250 .    
#> 1_115 0.1250 0.1250 .      .      .    .      .       .       .       .    
#> 1_116 0.1250 0.1250 .      .      .    .      .       .       .       .    
#> 1_117 .      .      .      .      .    .      .       .       .       .    
#> 1_118 0.1250 0.1250 .      .      .    .      .       .       .       .    
#> 1_119 0.2500 0.2500 .      .      .    .      .       .       .       .    
#> 1_120 0.2500 0.2500 .      .      .    .      .       .       .       .    
#> 1_121 .      .      0.0625 0.0625 .    0.0625 0.03125 0.03125 0.03125 .    
#> 1_122 .      .      0.0625 0.0625 .    0.0625 0.03125 0.03125 0.03125 .    
#> 1_123 .      .      0.0625 0.0625 .    0.0625 0.03125 0.03125 0.03125 .    
#> 1_124 .      .      0.0625 0.0625 .    0.0625 0.03125 0.03125 0.03125 .    
#> 1_125 0.0625 0.0625 0.0625 0.0625 .    0.0625 0.03125 0.03125 0.03125 .    
#> 1_126 0.0625 0.0625 0.0625 0.0625 .    0.0625 0.03125 0.03125 0.03125 .    
#> 1_127 0.0625 0.0625 0.0625 0.0625 .    0.0625 0.03125 0.03125 0.03125 .    
#> 1_128 0.0625 0.0625 0.0625 0.0625 .    0.0625 0.03125 0.03125 0.03125 .    
#> 1_129 0.0625 0.0625 .      .      .    .      .       .       .       .    
#> 1_130 0.2500 0.2500 .      .      .    .      .       .       .       .    
#> 1_131 0.2500 0.2500 .      .      .    .      .       .       .       .    
#> 1_132 0.2500 0.2500 .      .      .    .      .       .       .       .    
#> 1_133 0.5000 0.2500 .      .      .    .      .       .       .       .    
#> 1_134 0.2500 0.5000 .      .      .    .      .       .       .       .    
#> 1_135 .      .      0.5000 .      .    0.2500 0.12500 0.12500 0.12500 .    
#> 1_136 .      .      .      0.5000 .    0.2500 0.12500 0.12500 0.12500 .    
#> 1_137 .      .      .      .      0.50 .      0.25000 0.25000 0.25000 .    
#> 1_138 .      .      0.2500 0.2500 .    0.5000 0.25000 0.25000 0.25000 .    
#> 1_139 .      .      0.1250 0.1250 0.25 0.2500 0.50000 0.25000 0.25000 .    
#> 1_140 .      .      0.1250 0.1250 0.25 0.2500 0.25000 0.50000 0.25000 .    
#> 1_141 .      .      0.1250 0.1250 0.25 0.2500 0.25000 0.25000 0.50000 .    
#> 2_201 .      .      .      .      .    .      .       .       .       0.500
#> 2_202 .      .      .      .      .    .      .       .       .       .    
#> 2_203 .      .      .      .      .    .      .       .       .       .    
#> 2_204 .      .      .      .      .    .      .       .       .       0.250
#> 2_205 .      .      .      .      .    .      .       .       .       0.250
#> 2_206 .      .      .      .      .    .      .       .       .       0.250
#> 2_207 .      .      .      .      .    .      .       .       .       0.250
#> 2_208 .      .      .      .      .    .      .       .       .       0.250
#> 2_209 .      .      .      .      .    .      .       .       .       .    
#> 2_210 .      .      .      .      .    .      .       .       .       0.125
#> 2_211 .      .      .      .      .    .      .       .       .       0.125
#> 2_212 .      .      .      .      .    .      .       .       .       0.125
#> 2_213 .      .      .      .      .    .      .       .       .       0.125
#> 2_214 .      .      .      .      .    .      .       .       .       0.125
#>                                                                                
#> 1_101 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_102 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_103 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_104 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_105 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_106 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_107 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_108 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_109 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_110 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_111 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_112 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_113 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_114 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_115 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_116 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_117 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_118 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_119 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_120 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_121 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_122 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_123 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_124 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_125 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_126 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_127 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_128 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_129 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_130 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_131 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_132 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_133 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_134 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_135 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_136 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_137 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_138 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_139 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_140 .     .    .     .     .     .     .     .    .      .      .      .     
#> 1_141 .     .    .     .     .     .     .     .    .      .      .      .     
#> 2_201 .     .    0.250 0.250 0.250 0.250 0.250 .    0.1250 0.1250 0.1250 0.1250
#> 2_202 0.500 .    0.250 0.250 0.250 0.250 0.250 .    0.1250 0.1250 0.1250 0.1250
#> 2_203 .     0.50 .     .     .     .     .     .    0.2500 0.2500 .      .     
#> 2_204 0.250 .    0.500 0.250 0.250 0.250 0.250 .    0.2500 0.2500 0.1250 0.1250
#> 2_205 0.250 .    0.250 0.500 0.250 0.250 0.250 .    0.1250 0.1250 0.1250 0.1250
#> 2_206 0.250 .    0.250 0.250 0.500 0.250 0.250 .    0.1250 0.1250 0.1250 0.1250
#> 2_207 0.250 .    0.250 0.250 0.250 0.500 0.250 .    0.1250 0.1250 0.1250 0.1250
#> 2_208 0.250 .    0.250 0.250 0.250 0.250 0.500 .    0.1250 0.1250 0.2500 0.2500
#> 2_209 .     .    .     .     .     .     .     0.50 .      .      0.2500 0.2500
#> 2_210 0.125 0.25 0.250 0.125 0.125 0.125 0.125 .    0.5000 0.2500 0.0625 0.0625
#> 2_211 0.125 0.25 0.250 0.125 0.125 0.125 0.125 .    0.2500 0.5000 0.0625 0.0625
#> 2_212 0.125 .    0.125 0.125 0.125 0.125 0.250 0.25 0.0625 0.0625 0.5000 0.2500
#> 2_213 0.125 .    0.125 0.125 0.125 0.125 0.250 0.25 0.0625 0.0625 0.2500 0.5000
#> 2_214 0.125 .    0.125 0.125 0.125 0.125 0.250 0.25 0.0625 0.0625 0.2500 0.2500
#>             
#> 1_101 .     
#> 1_102 .     
#> 1_103 .     
#> 1_104 .     
#> 1_105 .     
#> 1_106 .     
#> 1_107 .     
#> 1_108 .     
#> 1_109 .     
#> 1_110 .     
#> 1_111 .     
#> 1_112 .     
#> 1_113 .     
#> 1_114 .     
#> 1_115 .     
#> 1_116 .     
#> 1_117 .     
#> 1_118 .     
#> 1_119 .     
#> 1_120 .     
#> 1_121 .     
#> 1_122 .     
#> 1_123 .     
#> 1_124 .     
#> 1_125 .     
#> 1_126 .     
#> 1_127 .     
#> 1_128 .     
#> 1_129 .     
#> 1_130 .     
#> 1_131 .     
#> 1_132 .     
#> 1_133 .     
#> 1_134 .     
#> 1_135 .     
#> 1_136 .     
#> 1_137 .     
#> 1_138 .     
#> 1_139 .     
#> 1_140 .     
#> 1_141 .     
#> 2_201 0.1250
#> 2_202 0.1250
#> 2_203 .     
#> 2_204 0.1250
#> 2_205 0.1250
#> 2_206 0.1250
#> 2_207 0.1250
#> 2_208 0.2500
#> 2_209 0.2500
#> 2_210 0.0625
#> 2_211 0.0625
#> 2_212 0.2500
#> 2_213 0.2500
#> 2_214 0.5000