Normalise a Ped object dataframe

Normalise dataframe for a Ped object

Usage

norm_ped(
  ped_df,
  na_strings = c("NA", ""),
  missid = NA_character_,
  try_num = FALSE,
  cols_used_del = FALSE
)

Arguments

ped_df

A data.frame with the individuals informations. The minimum columns required are:

id individual identifiers

dadid biological fathers identifiers

momid biological mothers identifiers

sex of the individual

famid family identifiers

The famid column, if provided, will be merged to the ids field separated by an underscore using the upd_famid() function.

The following columns are also recognize and will be transformed with the vect_to_binary() function:

avail status -> is the individual available

deceased status -> is the individual dead

The values recognized for those columns are 1 or 0, TRUE or FALSE.

The fertility column will be transformed to an ordered factor using the fertility_to_factor() function. infertile_choice_na < infertile < fertile The following values are recognized:

character() or factor() : "fertile", "infertile", "steril", "infertile_choice", "infertile_na", "infertile_choice_na"

numeric() : 1 = "fertile", 0 = "infertile"

logical() : TRUE = "fertile", FALSE = "infertile"

na_strings

Vector of strings to be considered as NA values.

missid

A character vector with the missing values identifiers. All the id, dadid and momid corresponding to those values will be set to NA_character_.

try_num

Boolean defining if the function should try to convert all the columns to numeric.

cols_used_del

Boolean defining if the columns that will be used should be deleted.

Value

A dataframe with different variable correctly standardized and with the errors identified in the error column

Details

Normalise a dataframe and check for columns correspondance to be able to use it as an input to create a Ped object. Multiple test are done and errors are checked.

Will be considered available any individual with no 'NA' values in the available column. Duplicated id will nullify the relationship of the individual. All individuals with errors will be remove from the dataframe and will be transfered to the error dataframe.

A number of checks are done to ensure the dataframe is correct:

On identifiers:

All ids (id, dadid, momid, famid) are not empty (!= "")

All id are unique (no duplicated)

All dadid and momid are unique in the id column (no duplicated)

id is not the same as dadid or momid

Either have both parents or none

On sex:

All sex code are either male, female, or unknown.

No parents are infertile.

All fathers are male

All mothers are female

See also

Examples

df <- data.frame(
    id = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10),
    dadid = c("A", 0, 1, 3, 0, 4, 1, 0, 6, 6),
    momid = c(0, 0, 2, 2, 0, 5, 2, 0, 8, 8),
    sex = c(1, 2, "m", "man", "f", "male", "m", 3, NA, "f"),
    avail = c("A", "1", 0, NA, 1, 0, 1, 0, 1, 0),
    famid = c(1, 1, 1, 1, 1, 1, 1, 2, 2, 2),
    fertility = c(
      "TRUE", "FALSE", TRUE, FALSE, 1, 0, "fertile", "infertile", 1, "TRUE"
    ), deceased = c("TRUE", "FALSE", TRUE, FALSE, 1, 0, 1, 0, 1, 0)
)
tryCatch(
    norm_ped(df),
    error = function(e) print(e)
)
#> Warning: NAs introduced by coercion
#>      id dadid momid     sex avail famid fertility deceased
#> 1   1_1   1_A   1_0    male    NA     1   fertile     TRUE
#> 2   1_2   1_0   1_0  female  TRUE     1 infertile    FALSE
#> 3   1_3   1_1   1_2    male FALSE     1   fertile     TRUE
#> 4   1_4   1_3   1_2    male    NA     1 infertile    FALSE
#> 5   1_5   1_0   1_0  female  TRUE     1   fertile     TRUE
#> 6   1_6   1_4   1_5    male FALSE     1 infertile    FALSE
#> 7   1_7   1_1   1_2    male  TRUE     1   fertile     TRUE
#> 8   2_8   2_0   2_0  female FALSE     2 infertile    FALSE
#> 9   2_9   2_6   2_8 unknown  TRUE     2   fertile     TRUE
#> 10 2_10   2_6   2_8  female FALSE     2   fertile    FALSE
#>                         error
#> 1                        <NA>
#> 2  is-infertile-but-is-parent
#> 3                        <NA>
#> 4  is-infertile-but-is-parent
#> 5                        <NA>
#> 6                        <NA>
#> 7                        <NA>
#> 8  is-infertile-but-is-parent
#> 9                        <NA>
#> 10                       <NA>